Assort ViTs in torch-em (#7)
